VZ-9633.  Refactor Cert-Manager component names to more accurately reflect what they are - Update other references where needed
