+++ platform-operator/controllers/verrazzano/component/certmanager/common/acme_utils.go
@@ -4,9 +4,10 @@
 package common
 
 import (
+	"strings"
+
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
-	"strings"
 )
 
 func IsLetsEncryptProductionEnv(acme interface{}) bool {
