+++ platform-operator/controllers/verrazzano/component/certmanager/common/acme_utils_test.go
@@ -4,10 +4,11 @@
 package common
 
 import (
+	"testing"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
-	"testing"
 )
 
 // TestIsLetsEncryptProductionEnv tests the IsLetsEncryptProductionEnv functions
