+++ platform-operator/controllers/verrazzano/component/certmanager/common/config.go
@@ -5,6 +5,7 @@ package common
 
 import (
 	"context"
+
 	"github.com/verrazzano/verrazzano/pkg/k8sutil"
 	"github.com/verrazzano/verrazzano/pkg/log/vzlog"
 	"github.com/verrazzano/verrazzano/pkg/vzcr"
