+++ platform-operator/controllers/verrazzano/component/certmanager/common/config_test.go
@@ -4,6 +4,8 @@
 package common
 
 import (
+	"testing"
+
 	acmev1 "github.com/cert-manager/cert-manager/pkg/apis/acme/v1"
 	certv1 "github.com/cert-manager/cert-manager/pkg/apis/certmanager/v1"
 	"github.com/stretchr/testify/assert"
@@ -20,7 +22,6 @@ import (
 	k8scheme "k8s.io/client-go/kubernetes/scheme"
 	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"testing"
 )
 
 const (
